+++ Glyssen/BlockNavigator.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Glyssen.Shared.Bundle;
 using SIL.Extensions;
 using SIL.Scripture;
