+++ Glyssen/Controls/CastSizePlanningOptions.cs
@@ -3,7 +3,6 @@ using System.Drawing;
 using System.Windows.Forms;
 using Glyssen.Bundle;
 using Glyssen.Dialogs;
-using Glyssen.Shared.Bundle;
 using Glyssen.Utilities;
 
 namespace Glyssen.Controls
