+++ Glyssen/Dialogs/ProjectSettingsViewModel.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Linq;
 using Glyssen.Bundle;
-using Glyssen.Shared.Bundle;
 using SIL.Scripture;
 using SIL.Windows.Forms.WritingSystems;
