+++ Glyssen/ReferenceTextProxy.cs
@@ -13,7 +13,12 @@ using Glyssen.Shared.Bundle;
 
 namespace Glyssen
 {
-	public class ReferenceTextIdentifier : IReferenceTextIdentifier
+	/// <summary>
+	/// Lets us work with a reference text even when we don't have an actual reference text object.
+	/// Originally (maybe still?) this enabled the use of "proprietary" reference texts
+	/// which can't ship with Glyssen for IP reasons but which we know FCBH is using.
+	/// </summary>
+	public class ReferenceTextProxy : IReferenceTextProxy
 	{
 		private const string kDistFilesReferenceTextDirectoryName = "reference_texts";
 
@@ -38,7 +43,7 @@ namespace Glyssen
 		internal static Action<Exception, string, string> ErrorReporterForCopyrightedReferenceTexts { get; set; }
 		#endregion
 
-		private static List<ReferenceTextIdentifier> s_allAvailable;
+		private static List<ReferenceTextProxy> s_allAvailable;
 		private static bool s_allAvailableLoaded = false;
 		private static string s_proprietaryReferenceTextProjectFileLocation;
 
@@ -61,7 +66,7 @@ namespace Glyssen
 			}
 		}
 
-		private ReferenceTextIdentifier(ReferenceTextType type, GlyssenDblTextMetadata metadata = null)
+		private ReferenceTextProxy(ReferenceTextType type, GlyssenDblTextMetadata metadata = null)
 		{
 			Debug.Assert(IsStandardReferenceText(type));
 			m_referenceTextType = type;
@@ -69,7 +74,7 @@ namespace Glyssen
 			m_metadata = metadata ?? LoadMetadata(type);
 		}
 
-		private ReferenceTextIdentifier(ReferenceTextType type, string customId, GlyssenDblTextMetadata metadata)
+		private ReferenceTextProxy(ReferenceTextType type, string customId, GlyssenDblTextMetadata metadata)
 		{
 			Debug.Assert(!IsStandardReferenceText(type));
 			Debug.Assert(customId != null);
@@ -94,7 +99,7 @@ namespace Glyssen
 			}
 		}
 
-		public static IEnumerable<ReferenceTextIdentifier> AllAvailable
+		public static IEnumerable<ReferenceTextProxy> AllAvailable
 		{
 			get
 			{
@@ -113,27 +118,27 @@ namespace Glyssen
 			s_allAvailableLoaded = false;
 		}
 
-		public static ReferenceTextIdentifier GetOrCreate(ReferenceTextType referenceTextType, string proprietaryReferenceTextIdentifier = null)
+		public static ReferenceTextProxy GetOrCreate(ReferenceTextType referenceTextType, string proprietaryReferenceTextIdentifier = null)
 		{
-			ReferenceTextIdentifier identifier;
+			ReferenceTextProxy proxy;
 			bool standard = IsStandardReferenceText(referenceTextType);
 			if (s_allAvailable == null)
 			{
-				s_allAvailable = new List<ReferenceTextIdentifier>();
-				identifier = null;
+				s_allAvailable = new List<ReferenceTextProxy>();
+				proxy = null;
 			}
 			else
 			{
-				identifier = standard ? s_allAvailable.SingleOrDefault(i => i.Type == referenceTextType) :
+				proxy = standard ? s_allAvailable.SingleOrDefault(i => i.Type == referenceTextType) :
 					s_allAvailable.SingleOrDefault(i => i.CustomIdentifier == proprietaryReferenceTextIdentifier);
 			}
-			if (identifier == null)
+			if (proxy == null)
 			{
-				identifier = standard ? new ReferenceTextIdentifier(referenceTextType) :
-					new ReferenceTextIdentifier(referenceTextType, proprietaryReferenceTextIdentifier, null);
-				s_allAvailable.Add(identifier);
+				proxy = standard ? new ReferenceTextProxy(referenceTextType) :
+					new ReferenceTextProxy(referenceTextType, proprietaryReferenceTextIdentifier, null);
+				s_allAvailable.Add(proxy);
 			}
-			return identifier;
+			return proxy;
 		}
 
 		//public override bool Equals(object obj)
@@ -141,7 +146,7 @@ namespace Glyssen
 		//	return base.Equals(obj);
 		//}
 
-		//protected bool Equals(ReferenceTextIdentifier other)
+		//protected bool Equals(ReferenceTextProxy other)
 		//{
 		//	return m_referenceTextType == other.m_referenceTextType && Equals(m_metadata, other.m_metadata);
 		//}
@@ -154,12 +159,12 @@ namespace Glyssen
 		//	}
 		//}
 
-		//public static bool operator ==(ReferenceTextIdentifier left, ReferenceTextIdentifier right)
+		//public static bool operator ==(ReferenceTextProxy left, ReferenceTextProxy right)
 		//{
 		//	return Equals(left, right);
 		//}
 
-		//public static bool operator !=(ReferenceTextIdentifier left, ReferenceTextIdentifier right)
+		//public static bool operator !=(ReferenceTextProxy left, ReferenceTextProxy right)
 		//{
 		//	return !Equals(left, right);
 		//}
@@ -172,7 +177,7 @@ namespace Glyssen
 		private static void LoadAllAvailable()
 		{
 			if (s_allAvailable == null)
-				s_allAvailable = new List<ReferenceTextIdentifier>();
+				s_allAvailable = new List<ReferenceTextProxy>();
 			Tuple<Exception, string, string> firstLoadError = null;
 			var additionalErrors = new List<string>();
 			Action<Exception, string, string> errorReporter = (exception, token, path) =>
@@ -188,7 +193,7 @@ namespace Glyssen
 			{
 				var metadata = LoadMetadata(itm, errorReporter);
 				if (metadata != null)
-					s_allAvailable.Add(new ReferenceTextIdentifier(itm, metadata));
+					s_allAvailable.Add(new ReferenceTextProxy(itm, metadata));
 			}
 
 			if (ErrorReporterForCopyrightedReferenceTexts == null)
@@ -246,7 +251,7 @@ namespace Glyssen
 				ErrorReporterForCopyrightedReferenceTexts);
 			if (metadata != null)
 			{
-				s_allAvailable.Add(new ReferenceTextIdentifier(ReferenceTextType.Custom, customId, metadata));
+				s_allAvailable.Add(new ReferenceTextProxy(ReferenceTextType.Custom, customId, metadata));
 				return true;
 			}
 			return false;
@@ -286,7 +291,7 @@ namespace Glyssen
 				return false;
 
 			if (s_allAvailable == null)
-				s_allAvailable = new List<ReferenceTextIdentifier>();
+				s_allAvailable = new List<ReferenceTextProxy>();
 			else if (IsCustomReferenceTextIdentifierInListOfAvailable(customId))
 				return true;
