+++ GlyssenShared/ReferenceTextType.cs
@@ -1,5 +1,9 @@
 ﻿namespace Glyssen.Shared
 {
+	/// <summary>
+	/// The reference text type.
+	/// Basically, is it a known, shipped language? Or something else?
+	/// </summary>
 	public enum ReferenceTextType
 	{
 		Unknown,
