+++ GlyssenShared/Script/GlyssenScript.cs
@@ -1,10 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Xml.Serialization;
 using Glyssen.Shared.Bundle;
 
 namespace Glyssen.Shared.Script
 {
+	/// <summary>
+	/// Defines a Glyssen script which can be exported and used by other applications, e.g. HearThis.
+	/// It seems odd that it extends something called Metadata, but that is basically done for serialization purposes.
+	/// Some of the DBL (Digital Bible Library) metadata is purposefully maintained from the bundle through the
+	/// project to the script (e.g. language info).
+	/// </summary>
 	[XmlRoot("glyssenscript")]
 	public class GlyssenScript : GlyssenDblTextMetadataBase
 	{
@@ -26,7 +31,7 @@ namespace Glyssen.Shared.Script
 		}
 
 		/// <summary>
-		/// This is used by consuming applications to know which versions it is capable of opening.
+		/// This is used by a consuming application to know which versions it is capable of opening.
 		/// We will use semver, so upgrade the minor for non-breaking changes and the major for
 		/// any change which could cause a consumer a problem when reading.
 		/// </summary>
@@ -43,8 +48,14 @@ namespace Glyssen.Shared.Script
 		public List<ScriptBook> Books { get; set; }
 	}
 
+	/// <summary>
+	/// Book data in the script
+	/// </summary>
 	public class ScriptBook
 	{
+		/// <summary>
+		/// The book ID, e.g. 'MAT'
+		/// </summary>
 		[XmlAttribute("id")]
 		public string Id { get; set; }
 
@@ -52,8 +63,15 @@ namespace Glyssen.Shared.Script
 		public List<ScriptChapter> Chapters { get; set; }
 	}
 
+	/// <summary>
+	/// Chapter data in the script
+	/// </summary>
 	public class ScriptChapter
 	{
+		/// <summary>
+		/// The chapter number.
+		/// Extra-biblical material (intro, book title, etc.) which comes before chapter 1 will be 0.
+		/// </summary>
 		[XmlAttribute("id")]
 		public int Id { get; set; }
 
@@ -61,40 +79,58 @@ namespace Glyssen.Shared.Script
 		public List<ScriptBlock> Blocks { get; set; }
 	}
 
+	/// <summary>
+	/// Block data in the script (essentially a line in the script)
+	/// </summary>
 	public class ScriptBlock
 	{
+		/// <summary>
+		/// A sequential number starting with 1 at the beginning of each chapter
+		/// </summary>
 		[XmlAttribute("id")]
 		public int Id { get; set; }
+
+		/// <summary>
+		/// The actor who speaks the line
+		/// </summary>
 		[XmlAttribute("actor")]
 		public string Actor { get; set; }
+
+		/// <summary>
+		/// The original sfm marker which applies to this text, e.g. 'm'
+		/// </summary>
 		[XmlAttribute("tag")]
 		public string Tag { get; set; }
+
+		/// <summary>
+		/// The applicable verse number or bridge
+		/// </summary>
 		[XmlAttribute("verse")]
 		public string Verse { get; set; }
+
+		/// <summary>
+		/// The Biblical character who speaks the line
+		/// </summary>
 		[XmlAttribute("character")]
 		public string Character { get; set; }
+
+		/// <summary>
+		/// How the line should be delivered, e.g. 'questioning'
+		/// </summary>
 		[XmlAttribute("delivery")]
 		public string Delivery { get; set; }
-		[XmlAttribute("file")]
-		public string File { get; set; }
 
 		[XmlElement("vern")]
-		public Vernacular Vernacular { get; set; }
+		public TextWithLanguage VernacularText { get; set; }
 
 		[XmlElement("primaryref")]
-		public Reference Primary { get; set; }
+		public TextWithLanguage PrimaryReferenceTextText { get; set; }
 
 		[XmlElement("secondaryref")]
-		public Reference Secondary { get; set; }
+		public TextWithLanguage SecondaryReferenceTextText { get; set; }
 	}
 
-	public class Vernacular
-	{
-		[XmlText]
-		public string Text { get; set; }
-	}
-
-	public class Reference
+	public class TextWithLanguage
 	{
 		[XmlAttribute("xml:lang", DataType = "language")]
 		public string LanguageCode { get; set; }
