+++ GlyssenTests/BlockNavigatorTests.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Glyssen;
-using Glyssen.Shared.Bundle;
 using NUnit.Framework;
 using SIL.Scripture;
