+++ GlyssenTests/GlyssenTests.csproj
@@ -105,6 +105,7 @@
     <Compile Include="BlockElementTests.cs" />
     <Compile Include="BlockMatchupTests.cs" />
     <Compile Include="BlockNavigatorTests.cs" />
+    <Compile Include="Dialogs\CastSizeCalculationUtilities.cs" />
     <Compile Include="ProjectCharacterDetailDataTests.cs" />
     <Compile Include="BlockTests.cs" />
     <Compile Include="BookMetadataTests.cs" />
