+++ GlyssenTests/Rules/CharacterGroupGeneratorTests.cs
@@ -8,7 +8,6 @@ using Glyssen.Bundle;
 using Glyssen.Character;
 using Glyssen.Dialogs;
 using Glyssen.Rules;
-using Glyssen.Shared.Bundle;
 using Glyssen.VoiceActor;
 using GlyssenTests.Properties;
 using NUnit.Framework;
