+++ GlyssenTests/Rules/CharacterGroupsAdjusterTests.cs
@@ -2,7 +2,6 @@
 using Glyssen.Bundle;
 using Glyssen.Character;
 using Glyssen.Rules;
-using Glyssen.Shared.Bundle;
 using Glyssen.VoiceActor;
 using GlyssenTests.Properties;
 using NUnit.Framework;
