+++ GlyssenTests/Rules/ProximityTests.cs
@@ -4,7 +4,6 @@ using Glyssen;
 using Glyssen.Bundle;
 using Glyssen.Character;
 using Glyssen.Rules;
-using Glyssen.Shared.Bundle;
 using GlyssenTests.Properties;
 using NUnit.Framework;
