+++ GlyssenTests/Rules/TrialGroupConfigurationTests.cs
@@ -5,7 +5,6 @@ using Glyssen;
 using Glyssen.Bundle;
 using Glyssen.Character;
 using Glyssen.Rules;
-using Glyssen.Shared.Bundle;
 using Glyssen.VoiceActor;
 using GlyssenTests.Properties;
 using NUnit.Framework;
