+++ RefTextDevUtilities/ReferenceTextData.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using Glyssen.Shared;
-using Glyssen.Shared.Bundle;
 
 namespace Glyssen.RefTextDevUtilities
 {
